refactoring for legacy headers
