+++ contrib/local-environment/oauth2-proxy.yaml
@@ -1,0 +1,27 @@
+server:
+  httpAddress: 0.0.0.0:4180
+  
+cookie:
+  secure: false
+  secret: OQINaROshtE9TcZkNAm-5Zs2Pv3xaWytBmc5W7sPX7w=
+  domains:
+    - .localtest.me
+
+providers:
+- id: dex
+  type: oidc
+  name: Dex
+  clientID: oauth2-proxy
+  clientSecret: b2F1dGgyLXByb3h5LWNsaWVudC1zZWNyZXQK
+  oidcConfig:
+    issuerURL: http://dex.localtest.me:4190/dex
+proxyOptions:
+  whitelistDomains:
+    - .localtest.me
+  emailDomains:
+    - .example.com
+upstreamConfig:
+  upstreams:
+  - id: httpbin
+    uri: http://httpbin.localtest.me:8080
+
