+++ pkg/apis/options/legacy_options.go
@@ -279,300 +279,6 @@ func (l *LegacyUpstreams) convert() (UpstreamConfig, error) {
 	return upstreams, nil
 }
 
-type LegacyHeaders struct {
-	PassBasicAuth     bool `flag:"pass-basic-auth" cfg:"pass_basic_auth"`
-	PassAccessToken   bool `flag:"pass-access-token" cfg:"pass_access_token"`
-	PassUserHeaders   bool `flag:"pass-user-headers" cfg:"pass_user_headers"`
-	PassAuthorization bool `flag:"pass-authorization-header" cfg:"pass_authorization_header"`
-
-	SetBasicAuth     bool `flag:"set-basic-auth" cfg:"set_basic_auth"`
-	SetXAuthRequest  bool `flag:"set-xauthrequest" cfg:"set_xauthrequest"`
-	SetAuthorization bool `flag:"set-authorization-header" cfg:"set_authorization_header"`
-
-	PreferEmailToUser    bool   `flag:"prefer-email-to-user" cfg:"prefer_email_to_user"`
-	BasicAuthPassword    string `flag:"basic-auth-password" cfg:"basic_auth_password"`
-	SkipAuthStripHeaders bool   `flag:"skip-auth-strip-headers" cfg:"skip_auth_strip_headers"`
-}
-
-func legacyHeadersFlagSet() *pflag.FlagSet {
-	flagSet := pflag.NewFlagSet("headers", pflag.ExitOnError)
-
-	flagSet.Bool("pass-basic-auth", true, "pass HTTP Basic Auth, X-Forwarded-User and X-Forwarded-Email information to upstream")
-	flagSet.Bool("pass-access-token", false, "pass OAuth access_token to upstream via X-Forwarded-Access-Token header")
-	flagSet.Bool("pass-user-headers", true, "pass X-Forwarded-User and X-Forwarded-Email information to upstream")
-	flagSet.Bool("pass-authorization-header", false, "pass the Authorization Header to upstream")
-
-	flagSet.Bool("set-basic-auth", false, "set HTTP Basic Auth information in response (useful in Nginx auth_request mode)")
-	flagSet.Bool("set-xauthrequest", false, "set X-Auth-Request-User and X-Auth-Request-Email response headers (useful in Nginx auth_request mode)")
-	flagSet.Bool("set-authorization-header", false, "set Authorization response headers (useful in Nginx auth_request mode)")
-
-	flagSet.Bool("prefer-email-to-user", false, "Prefer to use the Email address as the Username when passing information to upstream. Will only use Username if Email is unavailable, eg. htaccess authentication. Used in conjunction with -pass-basic-auth and -pass-user-headers")
-	flagSet.String("basic-auth-password", "", "the password to set when passing the HTTP Basic Auth header")
-	flagSet.Bool("skip-auth-strip-headers", true, "strips X-Forwarded-* style authentication headers & Authorization header if they would be set by oauth2-proxy")
-
-	return flagSet
-}
-
-// convert takes the legacy request/response headers and converts them to
-// the new format for InjectRequestHeaders and InjectResponseHeaders
-func (l *LegacyHeaders) convert() ([]Header, []Header) {
-	return l.getRequestHeaders(), l.getResponseHeaders()
-}
-
-func (l *LegacyHeaders) getRequestHeaders() []Header {
-	requestHeaders := []Header{}
-
-	if l.PassBasicAuth && l.BasicAuthPassword != "" {
-		requestHeaders = append(requestHeaders, getBasicAuthHeader(l.PreferEmailToUser, l.BasicAuthPassword))
-	}
-
-	// In the old implementation, PassUserHeaders is a subset of PassBasicAuth
-	if l.PassBasicAuth || l.PassUserHeaders {
-		requestHeaders = append(requestHeaders, getPassUserHeaders(l.PreferEmailToUser)...)
-		requestHeaders = append(requestHeaders, getPreferredUsernameHeader())
-	}
-
-	if l.PassAccessToken {
-		requestHeaders = append(requestHeaders, getPassAccessTokenHeader())
-	}
-
-	if l.PassAuthorization {
-		requestHeaders = append(requestHeaders, getAuthorizationHeader())
-	}
-
-	for i := range requestHeaders {
-		requestHeaders[i].PreserveRequestValue = !l.SkipAuthStripHeaders
-	}
-
-	return requestHeaders
-}
-
-func (l *LegacyHeaders) getResponseHeaders() []Header {
-	responseHeaders := []Header{}
-
-	if l.SetXAuthRequest {
-		responseHeaders = append(responseHeaders, getXAuthRequestHeaders()...)
-		if l.PassAccessToken {
-			responseHeaders = append(responseHeaders, getXAuthRequestAccessTokenHeader())
-		}
-	}
-
-	if l.SetBasicAuth {
-		responseHeaders = append(responseHeaders, getBasicAuthHeader(l.PreferEmailToUser, l.BasicAuthPassword))
-	}
-
-	if l.SetAuthorization {
-		responseHeaders = append(responseHeaders, getAuthorizationHeader())
-	}
-
-	return responseHeaders
-}
-
-func getBasicAuthHeader(preferEmailToUser bool, basicAuthPassword string) Header {
-	claim := "user"
-	if preferEmailToUser {
-		claim = "email"
-	}
-
-	return Header{
-		Name: "Authorization",
-		Values: []HeaderValue{
-			{
-				ClaimSource: &ClaimSource{
-					Claim:  claim,
-					Prefix: "Basic ",
-					BasicAuthPassword: &SecretSource{
-						Value: []byte(basicAuthPassword),
-					},
-				},
-			},
-		},
-	}
-}
-
-func getPassUserHeaders(preferEmailToUser bool) []Header {
-	headers := []Header{
-		{
-			Name: "X-Forwarded-Groups",
-			Values: []HeaderValue{
-				{
-					ClaimSource: &ClaimSource{
-						Claim: "groups",
-					},
-				},
-			},
-		},
-	}
-
-	if preferEmailToUser {
-		return append(headers,
-			Header{
-				Name: "X-Forwarded-User",
-				Values: []HeaderValue{
-					{
-						ClaimSource: &ClaimSource{
-							Claim: "email",
-						},
-					},
-				},
-			},
-		)
-	}
-
-	return append(headers,
-		Header{
-			Name: "X-Forwarded-User",
-			Values: []HeaderValue{
-				{
-					ClaimSource: &ClaimSource{
-						Claim: "user",
-					},
-				},
-			},
-		},
-		Header{
-			Name: "X-Forwarded-Email",
-			Values: []HeaderValue{
-				{
-					ClaimSource: &ClaimSource{
-						Claim: "email",
-					},
-				},
-			},
-		},
-	)
-}
-
-func getPassAccessTokenHeader() Header {
-	return Header{
-		Name: "X-Forwarded-Access-Token",
-		Values: []HeaderValue{
-			{
-				ClaimSource: &ClaimSource{
-					Claim: "access_token",
-				},
-			},
-		},
-	}
-}
-
-func getAuthorizationHeader() Header {
-	return Header{
-		Name: "Authorization",
-		Values: []HeaderValue{
-			{
-				ClaimSource: &ClaimSource{
-					Claim:  "id_token",
-					Prefix: "Bearer ",
-				},
-			},
-		},
-	}
-}
-
-func getPreferredUsernameHeader() Header {
-	return Header{
-		Name: "X-Forwarded-Preferred-Username",
-		Values: []HeaderValue{
-			{
-				ClaimSource: &ClaimSource{
-					Claim: "preferred_username",
-				},
-			},
-		},
-	}
-}
-
-func getXAuthRequestHeaders() []Header {
-	headers := []Header{
-		{
-			Name: "X-Auth-Request-User",
-			Values: []HeaderValue{
-				{
-					ClaimSource: &ClaimSource{
-						Claim: "user",
-					},
-				},
-			},
-		},
-		{
-			Name: "X-Auth-Request-Email",
-			Values: []HeaderValue{
-				{
-					ClaimSource: &ClaimSource{
-						Claim: "email",
-					},
-				},
-			},
-		},
-		{
-			Name: "X-Auth-Request-Preferred-Username",
-			Values: []HeaderValue{
-				{
-					ClaimSource: &ClaimSource{
-						Claim: "preferred_username",
-					},
-				},
-			},
-		},
-		{
-			Name: "X-Auth-Request-Groups",
-			Values: []HeaderValue{
-				{
-					ClaimSource: &ClaimSource{
-						Claim: "groups",
-					},
-				},
-			},
-		},
-	}
-
-	return headers
-}
-
-func getXAuthRequestAccessTokenHeader() Header {
-	return Header{
-		Name: "X-Auth-Request-Access-Token",
-		Values: []HeaderValue{
-			{
-				ClaimSource: &ClaimSource{
-					Claim: "access_token",
-				},
-			},
-		},
-	}
-}
-
-type LegacyServer struct {
-	MetricsAddress       string   `flag:"metrics-address" cfg:"metrics_address"`
-	MetricsSecureAddress string   `flag:"metrics-secure-address" cfg:"metrics_secure_address"`
-	MetricsTLSCertFile   string   `flag:"metrics-tls-cert-file" cfg:"metrics_tls_cert_file"`
-	MetricsTLSKeyFile    string   `flag:"metrics-tls-key-file" cfg:"metrics_tls_key_file"`
-	HTTPAddress          string   `flag:"http-address" cfg:"http_address"`
-	HTTPSAddress         string   `flag:"https-address" cfg:"https_address"`
-	TLSCertFile          string   `flag:"tls-cert-file" cfg:"tls_cert_file"`
-	TLSKeyFile           string   `flag:"tls-key-file" cfg:"tls_key_file"`
-	TLSMinVersion        string   `flag:"tls-min-version" cfg:"tls_min_version"`
-	TLSCipherSuites      []string `flag:"tls-cipher-suite" cfg:"tls_cipher_suites"`
-}
-
-func legacyServerFlagset() *pflag.FlagSet {
-	flagSet := pflag.NewFlagSet("server", pflag.ExitOnError)
-
-	flagSet.String("metrics-address", "", "the address /metrics will be served on (e.g. \":9100\")")
-	flagSet.String("metrics-secure-address", "", "the address /metrics will be served on for HTTPS clients (e.g. \":9100\")")
-	flagSet.String("metrics-tls-cert-file", "", "path to certificate file for secure metrics server")
-	flagSet.String("metrics-tls-key-file", "", "path to private key file for secure metrics server")
-	flagSet.String("http-address", "127.0.0.1:4180", "[http://]<addr>:<port> or unix://<path> to listen on for HTTP clients")
-	flagSet.String("https-address", ":443", "<addr>:<port> to listen on for HTTPS clients")
-	flagSet.String("tls-cert-file", "", "path to certificate file")
-	flagSet.String("tls-key-file", "", "path to private key file")
-	flagSet.String("tls-min-version", "", "minimal TLS version for HTTPS clients (either \"TLS1.2\" or \"TLS1.3\")")
-	flagSet.StringSlice("tls-cipher-suite", []string{}, "restricts TLS cipher suites to those listed (e.g. TLS_RSA_WITH_RC4_128_SHA) (may be given multiple times)")
-
-	return flagSet
-}
-
 type LegacyProvider struct {
 	ClientID         string `flag:"client-id" cfg:"client_id"`
 	ClientSecret     string `flag:"client-secret" cfg:"client_secret"`
@@ -705,50 +411,6 @@ func legacyGoogleFlagSet() *pflag.FlagSet {
 	flagSet.String("google-target-principal", "", "the target principal to impersonate when using ADC")
 
 	return flagSet
-}
-
-func (l LegacyServer) convert() (Server, Server) {
-	appServer := Server{
-		BindAddress:       l.HTTPAddress,
-		SecureBindAddress: l.HTTPSAddress,
-	}
-	if l.TLSKeyFile != "" || l.TLSCertFile != "" {
-		appServer.TLS = &TLS{
-			Key: &SecretSource{
-				FromFile: l.TLSKeyFile,
-			},
-			Cert: &SecretSource{
-				FromFile: l.TLSCertFile,
-			},
-			MinVersion: l.TLSMinVersion,
-		}
-		if len(l.TLSCipherSuites) != 0 {
-			appServer.TLS.CipherSuites = l.TLSCipherSuites
-		}
-		// Preserve backwards compatibility, only run one server
-		appServer.BindAddress = ""
-	} else {
-		// Disable the HTTPS server if there's no certificates.
-		// This preserves backwards compatibility.
-		appServer.SecureBindAddress = ""
-	}
-
-	metricsServer := Server{
-		BindAddress:       l.MetricsAddress,
-		SecureBindAddress: l.MetricsSecureAddress,
-	}
-	if l.MetricsTLSKeyFile != "" || l.MetricsTLSCertFile != "" {
-		metricsServer.TLS = &TLS{
-			Key: &SecretSource{
-				FromFile: l.MetricsTLSKeyFile,
-			},
-			Cert: &SecretSource{
-				FromFile: l.MetricsTLSCertFile,
-			},
-		}
-	}
-
-	return appServer, metricsServer
 }
 
 func (l *LegacyProvider) convert() (Providers, error) {
